backend/requirements.txt
asgiref==3.8.1
black==25.1.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
Django==5.2.3
django-sh==1.0.27
django-tenants==3.8.0
djangorestframework==3.16.0
idna==3.10
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
psycopg2-binary==2.9.10
python-decouple==3.8
requests==2.32.4
sqlparse==0.5.3
urllib3==2.5.0
python-dateutil==2.9.0.post0
djangorestframework_simplejwt==5.5.1
stripe==12.3.0